updating unit test
The unit test clearly shows that the  energy driver does not work at
all. The reason is due to the fact that the initial energy is set to nan
and thus does not allow the sum up of the forthcoming calibrated
energies. Setting the value to nan is good but given there are some
associated calorimeter hits, the initial value must be set to 0.0 at the
beginning of the loop.