Replace map by vector of pair
Previously, the map of cuts was indexed over the measurement name which
was a very bad idea since two or more cuts might rely to the same
measurement : consequently, when selecting internal TOF and external
TOF, only one of the two was used ! We now use a vector of pair value
avoiding the removing of duplicate measurement label.

Another issue was the discoupling with the cut manager since cut entries
were copied by value and not by reference. The initialization of cuts
and thus the creation of them is done within the initialization process
of the channel cut.